Add stupid-indent-mode as indent function handler
The ident function included in fsharp-mode to my knowledge is based in large on
the smie implementation for sml-mode. This often does the right thing but does
not adhere to the default F# style guide. Giving the user the option to use a
simple indentation like in other editors like Sublime can make their life a bit
easier.